back/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2
